search/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
newspaper3k==0.2.8
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4